refactor: WorkManager 관련 클래스 `checker` 패키지로 이동 및 `ProgressCheckWorker` 이름 변경
