+++ android/app/src/main/java/com/mulkkam/data/checker/ProgressCheckerImpl.kt
@@ -1,15 +1,16 @@
-package com.mulkkam.data.work
+package com.mulkkam.data.checker
 
 import androidx.work.OneTimeWorkRequestBuilder
 import androidx.work.WorkManager
-import com.mulkkam.domain.work.ProgressChecker
+import com.mulkkam.data.work.ProgressWorker
+import com.mulkkam.domain.checker.ProgressChecker
 import java.util.UUID
 
 class ProgressCheckerImpl(
     private val workManager: WorkManager,
 ) : ProgressChecker {
     override fun checkCurrentAchievementRate(): UUID {
-        val workRequest = OneTimeWorkRequestBuilder<ProgressCheckWorker>().build()
+        val workRequest = OneTimeWorkRequestBuilder<ProgressWorker>().build()
 
         workManager.enqueue(workRequest)
         return workRequest.id
