+++ android/app/src/main/java/com/mulkkam/di/WorkInjection.kt
@@ -3,30 +3,20 @@ package com.mulkkam.di
 import android.content.Context
 import androidx.work.Configuration
 import androidx.work.WorkManager
-import com.mulkkam.data.work.CalorieSchedulerImpl
-import com.mulkkam.data.work.ProgressCheckerImpl
 import com.mulkkam.data.work.WorkerFactory
 import com.mulkkam.di.RepositoryInjection.cupsRepository
 import com.mulkkam.di.RepositoryInjection.healthRepository
 import com.mulkkam.di.RepositoryInjection.intakeRepository
 import com.mulkkam.di.RepositoryInjection.membersRepository
 import com.mulkkam.di.RepositoryInjection.notificationRepository
-import com.mulkkam.domain.work.CalorieScheduler
-import com.mulkkam.domain.work.ProgressChecker
 
 object WorkInjection {
-    private lateinit var workManager: WorkManager
-
-    val calorieScheduler: CalorieScheduler by lazy {
-        CalorieSchedulerImpl(workManager)
-    }
-
-    val progressChecker: ProgressChecker by lazy {
-        ProgressCheckerImpl(workManager)
-    }
+    private var _workManager: WorkManager? = null
+    val workManager: WorkManager
+        get() = _workManager ?: throw IllegalStateException()
 
     fun init(context: Context) {
-        if (::workManager.isInitialized.not()) {
+        if (_workManager == null) {
             val config =
                 Configuration
                     .Builder()
@@ -35,7 +25,7 @@ object WorkInjection {
                     ).build()
 
             WorkManager.initialize(context, config)
-            workManager = WorkManager.getInstance(context)
+            _workManager = WorkManager.getInstance(context)
         }
     }
 }
