+++ android/app/src/main/java/com/mulkkam/domain/checker/CalorieChecker.kt
@@ -1,0 +1,9 @@
+package com.mulkkam.domain.checker
+
+interface CalorieChecker {
+    fun checkCalorie(intervalHours: Long)
+
+    companion object {
+        const val DEFAULT_CHECK_CALORIE_INTERVAL_HOURS: Long = 2L
+    }
+}
